chore(trading): remove waits and tests
